\name{AstroEmpires-package}
\alias{AstroEmpires-package}
\alias{AstroEmpires}
\docType{package}
\title{
\packageTitle{AstroEmpires}
}
\description{
\packageDescription{AstroEmpires}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AstroEmpires}
\packageIndices{AstroEmpires}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AstroEmpires}

Maintainer: \packageMaintainer{AstroEmpires}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
