% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bases.R
\name{read.bases}
\alias{read.bases}
\title{read.bases}
\usage{
read.bases(input)
}
\arguments{
\item{input}{string vector from the base page.}
}
\value{
data frame of Terrain Type Position for bases
}
\description{
generates the bases table.
}
\examples{
# default
bases <- read.bases(input=stdin())

}
