% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.structures.R
\name{read.structures}
\alias{read.structures}
\title{read.structures}
\usage{
read.structures(input)
}
\arguments{
\item{input}{string vector from the structure page.}
}
\value{
data frame of base structure levels.
}
\description{
generates the structure table.
}
\examples{
# default
bases <- read.structures(input=stdin())

}
