% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.techlevels.R
\name{read.techlevels}
\alias{read.techlevels}
\title{read.techlevels}
\usage{
read.techlevels(input)
}
\arguments{
\item{input}{string vector from the technologies page.}
}
\value{
matrix of technology levels
}
\description{
generates the structure table.
}
\examples{
# default
bases <- read.techlevels(input=stdin())

}
